move test to new test_media.py
